+++ src/main/java/br/com/banco/ContaPoupanca.java
@@ -3,7 +3,7 @@ package br.com.banco;
 import java.math.BigDecimal;
 
 public class ContaPoupanca extends Conta {
-	public ContaPoupanca(Long numeroConta, BigDecimal saldo, Agencia agencia) {
+	public ContaPoupanca(int numeroConta, BigDecimal saldo, Agencia agencia) {
 		super(numeroConta, saldo, agencia);
 	}
 }
